Trade_Journal_Agent/pyproject.toml
[project]
name = "trade-journal-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic",
    "matplotlib",
    "pandas",
    "python-dotenv",
    "yfinance",
]
